+++ modules/apps/commerce/commerce-item-selector-api/src/main/java/com/liferay/commerce/item/selector/SimpleSiteItemSelectorCriterion.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.item.selector.criterion;
+package com.liferay.commerce.item.selector;
 
 import com.liferay.item.selector.BaseItemSelectorCriterion;
